+++ src/Allegiance.Blazor.Highcharts/Options/Condition.cs
@@ -5,14 +5,14 @@ namespace Allegiance.Blazor.Highcharts.Options
     public partial class Condition
     {
         [JsonProperty("maxWidth", NullValueHandling = NullValueHandling.Ignore)]
-        public long? MaxWidth { get; set; }
+        public int? MaxWidth { get; set; }
 
         public Condition()
         {
 
         }
 
-        public Condition(long? maxWidth = null)
+        public Condition(int? maxWidth = null)
         {
             MaxWidth = maxWidth;
         }
